dashboard integration clean up
