+++ lib/plugins/emit/index.js
@@ -5,7 +5,7 @@ const BbPromise = require('bluebird');
 const fdk = require('@serverless/fdk');
 const path = require('path');
 const stdin = require('get-stdin');
-const getAuthToken = require('../../utils/getAuthToken');
+const isLoggedIn = require('../../utils/isLoggedIn');
 const userStats = require('../../utils/userStats');
 const chalk = require('chalk');
 
@@ -120,8 +120,7 @@ class Emit {
   }
 
   emitEvent() {
-    const authToken = getAuthToken();
-    if (!authToken) {
+    if (!isLoggedIn()) {
       return BbPromise.reject(new this.serverless.classes
         .Error('Must be logged in to use this command. Please run "serverless login".'));
     }
