+++ lib/plugins/login/login.js
@@ -43,7 +43,6 @@ class Login {
         name: decoded.name,
         email: decoded.email,
         username: decoded.nickname,
-        auth: configuration.users[id].auth,
         dashboard: data,
       };
