+++ lib/plugins/platform/platform.js
@@ -7,7 +7,7 @@ const crypto = require('crypto');
 const _ = require('lodash');
 const platform = require('@serverless/platform-sdk');
 const getAccessKey = require('../../utils/getAccessKey');
-const configUtils = require('../../utils/config');
+const isLoggedIn = require('../../utils/isLoggedIn');
 
 class Platform {
   constructor(serverless, options) {
@@ -28,28 +28,20 @@ class Platform {
   }
 
   publishService() {
-    const config = configUtils.getConfig();
-    const currentId = config.userId;
-    const globalConfig = configUtils.getGlobalConfig();
-    let isLoggedIn = false;
-    if (globalConfig
-      && globalConfig.users
-      && globalConfig.users[currentId]
-      && globalConfig.users[currentId].dashboard) {
-      isLoggedIn = true;
+    if (!isLoggedIn()) {
+      return BbPromise.resolve();
     }
 
-    if (isLoggedIn) {
-      if (!this.config.tenant && !this.config.app) {
-        this.serverless.cli.log('WARNING: Missing "tenant" and "app" properties in serverless.yml');
-      } else if (this.config.tenant && !this.config.app) {
-        const errorMessage = ['Missing "app" property in serverless.yml'].join('');
-        throw new this.serverless.classes.Error(errorMessage);
-      } else if (!this.config.tenant && this.config.app) {
-        const errorMessage = ['Missing "tenant" property in serverless.yml'].join('');
-        throw new this.serverless.classes.Error(errorMessage);
-      }
+    if (!this.config.tenant && !this.config.app) {
+      this.serverless.cli.log('WARNING: Missing "tenant" and "app" properties in serverless.yml');
+    } else if (this.config.tenant && !this.config.app) {
+      const errorMessage = ['Missing "app" property in serverless.yml'].join('');
+      throw new this.serverless.classes.Error(errorMessage);
+    } else if (!this.config.tenant && this.config.app) {
+      const errorMessage = ['Missing "tenant" property in serverless.yml'].join('');
+      throw new this.serverless.classes.Error(errorMessage);
     }
+
     return getAccessKey(this.config.tenant).then(accessKey => {
       if (!accessKey || !this.config.app || !this.config.tenant) {
         return BbPromise.resolve();
@@ -127,7 +119,6 @@ class Platform {
             this.serverless.cli
               .log('Successfully published your service on the Serverless Platform');
             resolve();
-            process.exit(0);
           })
           .catch(err => {
             this.serverless.cli.log('Failed to publish your service on the Serverless Platform');
@@ -167,6 +158,20 @@ class Platform {
   }
 
   archiveService() {
+    if (!isLoggedIn()) {
+      return BbPromise.resolve();
+    }
+
+    if (!this.config.tenant && !this.config.app) {
+      this.serverless.cli.log('WARNING: Missing "tenant" and "app" properties in serverless.yml');
+    } else if (this.config.tenant && !this.config.app) {
+      const errorMessage = ['Missing "app" property in serverless.yml'].join('');
+      throw new this.serverless.classes.Error(errorMessage);
+    } else if (!this.config.tenant && this.config.app) {
+      const errorMessage = ['Missing "tenant" property in serverless.yml'].join('');
+      throw new this.serverless.classes.Error(errorMessage);
+    }
+
     return getAccessKey(this.config.tenant).then(accessKey => {
       if (!accessKey || !this.config.app || !this.config.tenant) {
         return BbPromise.resolve();
@@ -180,7 +185,7 @@ class Platform {
       return platform.archiveService(data)
         .then(() => {
           this.serverless.cli.log('Successfully archived your service on the Serverless Platform');
-          process.exit(0);
+          // process.exit(0);
         })
         .catch(err => {
           this.serverless.cli.log('Failed to archived your service on the Serverless Platform');
