+++ lib/plugins/run/index.js
@@ -20,7 +20,7 @@ const registerFunctionsToEventGateway = require('./utils/registerFunctionsToEven
 const manageLocalEmulator = require('./utils/manageLocalEmulator');
 const manageEventGateway = require('./utils/manageEventGateway');
 
-const getAuthToken = require('../../utils/getAuthToken');
+const isLoggedIn = require('../../utils/isLoggedIn');
 
 class Run {
   constructor(serverless, options) {
@@ -67,8 +67,7 @@ class Run {
     const EVENT_GATEWAY_VERSION = '0.5.15';
     const LOCAL_EMULATOR_VERSION = '0.1.18';
 
-    const authToken = getAuthToken();
-    if (!authToken) {
+    if (!isLoggedIn()) {
       throw new this.serverless.classes
         .Error('Must be logged in to use this command. Please run "serverless login".');
     }
